[SPARK-18219] Move commit protocol API from sql/core to core module
